server/requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
sqlparse==0.5.3
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
urllib3==2.3.0
